pathway apis added, need testing
