+++ apis/userApis.js
@@ -26,10 +26,11 @@ const userApis = {
     POST_session,
     POST_verifyEmail,
     POST_changePasswordForgot,
-    POST_changePassword,
     POST_forgotPassword,
+    POST_changePassword,
+    POST_changeSettings,
     GET_userByProfileUrl,
-    POST_login
+    POST_login,
 }
 
 
@@ -480,6 +481,84 @@ function POST_login(req, res) {
             }
         });
     }
+}
+
+
+function POST_changeSettings(req, res) {
+    const user = sanitize(req.body);
+    const password = user.password;
+
+    if (!user.password || !user.name || !user.email) {
+        console.log("Not all arguments provided for settings change.");
+        return res.status(400).send("No fields can be empty.");
+    }
+
+    const userQuery = {_id: user._id}
+
+    Users.findOne(userQuery, function(findUserErr, foundUser) {
+        // if error while trying to find current user
+        if (findUserErr) {
+            console.log("Error finding user in db when trying to update settings: ", findUserErr);
+            return res.status(500).send("Settings couldn't be updated. Try again later.");
+        }
+
+        if (!foundUser) {
+            console.log("Didn't find a user with given id when trying to update settings.");
+            return res.status(500).send("Settings couldn't be updated. Try again later.");
+        }
+
+        bcrypt.compare(password, foundUser.password, function (passwordError, passwordsMatch) {
+            // error comparing password to user's password, doesn't necessarily
+            // mean that the password is wrong
+            if (passwordError) {
+                console.log("Error comparing passwords when trying to update settings: ", passwordError);
+                return res.status(500).send("Settings couldn't be updated. Try again later.");
+            }
+
+            // user entered wrong password
+            if (!passwordsMatch) {
+                return res.status(400).send("Incorrect password");
+            }
+
+            // see if there's another user with the new email
+            const emailQuery = {email: user.email};
+            Users.findOne(emailQuery, function(emailQueryErr, userWithEmail) {
+                // don't want two users with the same email, so in case of db search
+                // failure, return unsuccessfully
+                if (emailQueryErr) {
+                    console.log("Error trying to find a user with the same email address as the one provided by user trying to change settings: ", emailQueryErr);
+                    return res.status(500).send("Settings couldn't be updated. Try again later.");
+                }
+
+                // someone else already has that email
+                if (userWithEmail && userWithEmail._id.toString() != foundUser._id.toString()) {
+                    return res.status(400).send("That email address is already taken.");
+                }
+
+                // all is good, update the user (as long as email and name are not blank)
+                if (user.email) {
+                    foundUser.email = user.email;
+                }
+                if (user.name) {
+                    foundUser.name = user.name;
+                }
+                if (typeof user.hideProfile === "boolean") {
+                    foundUser.hideProfile = user.hideProfile;
+                }
+
+                foundUser.save(function(saveErr, newUser) {
+                    // if there is an error saving the user's info
+                    if (saveErr) {
+                        console.log("Error when saving user's changed info: ", saveErr);
+                        return res.status(500).send("Settings couldn't be updated. Try again later.");
+                    }
+
+                    // settings change successful
+                    return res.json(newUser);
+                })
+            });
+        });
+    })
 }
