+++ src/components/pages/myPathways.js
@@ -43,7 +43,7 @@ class MyPathways extends Component {
                 else {
                     for (let i = 0; i < this.props.currentUser.pathways.length; i++) {
                         let id = this.props.currentUser.pathways[i].pathwayId;
-                        axios.get("/api/pathwayByIdNoContent", {
+                        axios.get("/api/pathway/pathwayByIdNoContent", {
                             params: {
                                 _id: id
                             }
@@ -104,7 +104,7 @@ class MyPathways extends Component {
             if (this.props.currentUser.completedPathways) {
                 for (let i = 0; i < this.props.currentUser.completedPathways.length; i++) {
                     let id = this.props.currentUser.completedPathways[i].pathwayId;
-                    axios.get("/api/pathwayByIdNoContent", {
+                    axios.get("/api/pathway/pathwayByIdNoContent", {
                         params: {
                             _id: id
                         }
